+++ backend/canisters/user/impl/src/updates/update_bot.rs
@@ -2,8 +2,7 @@ use crate::guards::caller_is_owner;
 use crate::{RuntimeState, execute_update};
 use canister_api_macros::update;
 use canister_tracing_macros::trace;
-use oc_error_codes::OCErrorCode;
-use types::{BotEvent, BotInstallationLocation, BotInstalledEvent, BotLifecycleEvent, BotNotification, OCResult};
+use types::OCResult;
 use user_canister::update_bot::*;
 
 #[update(guard = "caller_is_owner", msgpack = true)]
@@ -13,31 +12,9 @@ fn update_bot(args: Args) -> Response {
 }
 
 fn update_bot_impl(args: Args, state: &mut RuntimeState) -> OCResult {
-    state.data.user.verify_not_suspended()?;
-
+    let my_user_id = state.env.canister_id().into();
     let now = state.env.now();
-
-    if !state.data.user.update_bot_permissions(
-        args.bot_id,
-        args.granted_permissions.clone(),
-        args.granted_autonomous_permissions.clone(),
-        now,
-    ) {
-        return Err(OCErrorCode::BotNotFound.into());
-    }
-
-    let installed_by = state.env.canister_id().into();
-
-    state.push_bot_notification(Some(BotNotification {
-        event: BotEvent::Lifecycle(BotLifecycleEvent::Installed(BotInstalledEvent {
-            installed_by,
-            location: BotInstallationLocation::User(installed_by.into()),
-            granted_command_permissions: args.granted_permissions,
-            granted_autonomous_permissions: args.granted_autonomous_permissions.unwrap_or_default(),
-        })),
-        recipients: vec![args.bot_id],
-        timestamp: now,
-    }));
-
+    let notification = user_core::updates::update_bot(&mut state.data.user, args, my_user_id, now)?;
+    state.push_bot_notification(Some(notification));
     Ok(())
 }